pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-output-quality-lab"
version = "0.1.0"
description = "A small lab for systematically exploring and evaluating LLM prompt variants and temperatures."
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "Your Name", email = "you@example.com" }
]
license = { file = "LICENSE" }
dependencies = [
  "openai>=1.6.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
  "rich>=13.0.0",
  "pytest>=7.0.0",
]

[project.optional-dependencies]
dev = []

[project.scripts]
qolab = "qolab.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
